polymarket-bot/requirements.txt
py-clob-client>=0.19.0
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.0.0
eth-account>=0.10.0
schedule>=1.2.0
loguru>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
